Add option to show overridden repos
The option '--show-overridden-repos' aims to show in a column
the repos providing a package which have been overridden
because of older NVR or DNF configuration (priority, cost, etc).

'None' is displayed when there are no overridden repos.
This option is not applicable for Koji tag as source, 'N/A' is
displayed in this case.

With this option it's be easy to show packages that RDO repos
provide which override base CentOS repos.

Change-Id: I1c662c9ab60b950f0f20c4e18e12cd79730f6992