Update pubkey fingerprint while sending new version
We add here a function which checks if global var
"%sources_gpg_sign" is up-to-date. If not, it will
replace the pubkey fingerprint by the right one.

Note 1: we rely on HTTP/SSL certificate chain of
tarballs.opendev.org.

Note 2: this operation occurs during the sending
of a new release by rdo-trunk, so CI jobs will be
triggered and the "real" GPG verification will be
done while building RPM package. This should avoid us
to trigger twice CI jobs, because of obsolete pubkey
fingerprint. But still, any pubkey fingerprint change
needs to be reviewed.

Note 3: the check/update works in best-effort, this
should not fail the operation if errors occur.

Change-Id: I7b05649c63e83488d990a7cf6b2b1ff697181040